Refactored Duke to extract UI output methods
